+++ src/dropdown.js
@@ -25,7 +25,7 @@
   //
   // element - Textarea or contenteditable element.
   function Dropdown(element, completer, option) {
-    this.$el       = Dropdown.createElement();
+    this.$el       = Dropdown.createElement(option);
     this.completer = completer;
     this.id        = completer.id + 'dropdown';
     this._data     = []; // zipped data.
@@ -45,12 +45,12 @@
     // Class methods
     // -------------
 
-    createElement: function () {
+    createElement: function (option) {
       return $('<ul class="dropdown-menu"></ul>').css({
         display: 'none',
         left: 0,
         position: 'absolute',
-        zIndex: '100'
+        zIndex: option.zIndex
       });
     }
   });